Updated to adjust to the recent update
* Updated README.md
* Added a bunch of XML comments to improve usability
* Added PhaseCountdowns Node
* Added "Position" to PlayerNode
* Added "EquipValue" to PlayerStateNode
* Changed how the GSL is shut down to hopefully fix an exception when
shutting down while processing a new GameState (not entirely fixed,
suggestions welcome)

Sorry for being inactive for such a long time, but I was dealing with
some serious health issues and simply couldn't find the time or
motivation to work on this project. Things are getting better now and I
will try to update this project more frequently in the future.